pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "~4.2"
crispy-bootstrap5 = "^0.7"
django-markdownify = "^0.9.3"
django-mathfilters = "^1.0.0"
django-mdeditor = "^0.1.20"
psycopg2-binary = "^2.9.6"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.3"
pydantic = "^2.7.3"
djangorestframework = "^3.15.2"
django-colorfield = "^0.13.0"
django-admin-autocomplete-filter = "^0.7.1"
setuptools = "^80.9.0"
petrovich = "^2.0.1"
pymorphy3 = "^2.0.4"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
django-stubs = "^5.2.7"
djangorestframework-stubs = "<3.16.3"
django-debug-toolbar = "^4.1.0"
mypy = "^1.8.0"
ruff = "^0.13"
pre-commit = "^3.8.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.14.0"
pytest-django = "^4.8.0"
beautifulsoup4 = "^4.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
warn_return_any = true
show_error_codes = true
follow_imports = "silent"
check_untyped_defs = true
warn_unused_ignores = true
no_implicit_optional = true
disallow_untyped_defs = true
strict = true
warn_redundant_casts = true
warn_unused_configs = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict_concatenate = true
exclude = [
    "^.*/migrations/.*\\.py$",
    "^manage\\.py$",
]

[tool.django-stubs]
django_settings_module = "MoiGoroda.settings"

[tool.ruff]
line-length = 100
exclude = [
    "*/migrations/*",
]

[tool.ruff.format]
quote-style = "single"

# PyTest
[tool.pytest.ini_options]
addopts = "--import-mode=importlib --strict-markers"
DJANGO_SETTINGS_MODULE = "MoiGoroda.settings"
python_files = ["test_*"]
python_functions = ["test_*"]
markers = [
    "unit: Быстрые unit-тесты без БД (моки, чистая логика, изолированные компоненты)",
    "integration: Интеграционные тесты с БД (тестирование одного компонента с реальной БД)",
    "e2e: End-to-end тесты (полные сценарии взаимодействия нескольких компонентов)",
    "slow: Медленные тесты (запускать только в CI/CD)",
]
